requirements.txt
flask
pytest
PyMongo
pyinstaller
python-dotenv